refactor to move things around to slightly more sane (?) locations
