+++ code/lambda-functions/waze-data-api/src/api-models/getJamSnapshotResponse.ts
@@ -1,0 +1,13 @@
+import * as entities from '../../../shared-lib/src/entities'
+
+export class JamSnapshot extends entities.JamBase<entities.Point> {
+    
+}
+
+export class GetJamsListSnapshotResult 
+{
+    jams : JamSnapshot[];
+    timeframeReturned: entities.Timeframe;
+    nextTimeframe: entities.Timeframe;
+    previousTimeframe: entities.Timeframe;
+}
