+++ code/lambda-functions/waze-data-api/src/db/getJamListSnapshotQueryBuilder.ts
@@ -3,18 +3,8 @@ import * as entities from '../../../shared-lib/src/entities'
 import moment = require("moment");
 import { getJamSnapshotRequestModel } from "../api-models/getJamSnapshotRequestModel";
 import { QueryResult } from 'pg';
+import { GetJamsListSnapshotResult, JamSnapshot } from "../api-models/getJamSnapshotResponse";
 
-export class JamSnapshot extends entities.JamBase<entities.Point> {
-    
-}
-
-export class GetJamsListSnapshotResult 
-{
-    jams : JamSnapshot[];
-    timeframeReturned: entities.Timeframe;
-    nextTimeframe: entities.Timeframe;
-    previousTimeframe: entities.Timeframe;
-}
 
 export function mapSnapshotResultFromQueryQueryResult(queryResponse : QueryResult) : GetJamsListSnapshotResult
 {
